requirements.txt
requests>=2.32.3
PyGithub>=2.6.1
pyyaml>=6.0.2
Jinja2>=3.1.6
